Elder things side-channel
Idea from Sage: what if players are elder things feeding dreams into these mages, and that's how we control them?

Primarily driven from the need to write orders to _someone_ -- dispatch is too strange.  No need for the dreams to be obscure, can still be just as clear.  Presents another bunch of great ideas for objectives and fiction, though.